requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cffi==1.11.4
chardet==3.0.4
discord.py==1.2.3
idna==2.8
multidict==4.0.0
pycparser==2.18
PyNaCl==1.0.1
six==1.11.0
websockets==6.0
yarl==1.3.0
